pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sessionclean"
version = "0.1.0"
description = "Review and clean files created during your Windows session before shutdown"
readme = "README.md"
license = "GPL-3.0-only"
requires-python = ">=3.10"
authors = [
    { name = "Cristian" },
]
keywords = ["cleanup", "session", "files", "windows", "disk", "organizer"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Filesystems",
    "Topic :: Utilities",
]
dependencies = [
    "watchdog>=4.0",
    "customtkinter>=5.2",
    "pystray>=0.19",
    "Pillow>=10.0",
    "pywin32>=306",
    "send2trash>=1.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
sessionclean = "sessionclean.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
